requirements.txt
asgiref==3.6.0
autopep8==2.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-import-export==3.0.2
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
